requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
colorama==0.4.6
Django==4.2.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
sqlparse==0.4.4
tzdata==2023.3
python-dotenv~=1.0.0
djangorestframework~=3.14.0